Load / save  project

+++ Core/Project/Project.h
@@ -11,7 +11,7 @@ namespace FirstYear::Core {
 struct MonthItem {
   std::optional<QPixmap> photo;
   int angle = 0;
-  int scale = 1;
+  double scale = 1;
   QPoint center_coordinates = {0, 0};
   std::optional<QString> text;
   QString filter_id;
